Read/write attitribute values
Implemented tag reading and writing.
Refactored read to perform a deep copy of the variant.  This way
the free of the c memory side can happen immediately, leaving the
caller with idiomatic zig allocator style memory cleanup.  This
costs us a memcopy but I have decided it is worth it in order to
keep the API idiomatic with Zig.  I suppose later, we would implement
a zero-copy method that uses pointers to the underlying c-allocated
memory with a helper method to free the c memory.  This feels
likes premature optimization right now, though.